Fixed DateRangePicker directive.
Now it properly keeps track of min and max values and corrects display to reflect it.

Also, when clearing back the max and min values to null, the daterangepicker javacsript reloads it's values back to the initial default, ensuring a clean reset after loading templates.

Hooked up the x placeholder in the template select2 dropdown to fire the clear event (same as the clear button)